requirements.txt
dj-database-url==0.5.0
Django==2.1.3
django-heroku==0.3.1
django-static-jquery3==3.3.1.1
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2-binary==2.7.5
pytz==2018.7
whitenoise==4.1
